requirements.txt
setuptools
django==5.2.3
djangorestframework==3.15.1
drf-yasg==1.21.7
Pillow==10.3.0
psycopg-binary
psycopg2-binary
python-dotenv
social-auth-app-django
django-cors-headers
channels
openai
gunicorn
uvicorn
aiohttp